Add endpoints for orders

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Get, Post, Body, Delete, Param } from '@nestjs/common';
+import { Controller, Get, Post, Put, Body, Delete, Param, ParseUUIDPipe, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
+
 
 @Controller('products')
 export class ProductsController {
@@ -11,17 +13,31 @@ export class ProductsController {
         return this.productsService.getAll();
     }
     @Get('/:id')
-    public getById(@Param('id') id: string) {
-        return this.productsService.getById(id);
+    public getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        const prod = this.productsService.getById(id);
+        if(!prod) throw new NotFoundException('Product not found');
     }
+
     @Delete('/:id')
-    public deleteById(@Param('id') id: string) {
+    public deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if(!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
         this.productsService.deleteById(id);
         return { success: true };
     }
     @Post('/')
-    create(@Body() productData: CreateProductDTO) {
+    public create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
     }
+    @Put('/:id')
+    public update(
+        @Param('id', new ParseUUIDPipe()) id: string, 
+        @Body() productData: UpdateProductDTO,
+    ) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
 
+        this.productsService.updateById(id, productData);
+        return { success: true };
+    }
 }
